RD-1222 Extract widget code to separate files
